+++ package/bin/webex_api_client.py
@@ -26,11 +26,15 @@ def paging_get_request_to_webex(
     client_secret,
     params,
     response_tag,
-    is_custom_endpoint=False
+    is_custom_endpoint=False,
+    webex_account_region="us_ca",
+    method = "GET",
+    payload=None
 ):
     results = []
     # set the page_size
     params["max"] = _MAX_PAGE_SIZE if not params.get("max") else params["max"]
+
 
     paging = True
     next_page_link = None
@@ -48,7 +52,10 @@ def paging_get_request_to_webex(
                 client_secret,
                 params,
                 next_page_link,
-                is_custom_endpoint=is_custom_endpoint
+                is_custom_endpoint=is_custom_endpoint,
+                webex_account_region=webex_account_region,
+                method = method,
+                payload = payload
             )
 
             if data is None or len(data)==0:
@@ -91,20 +98,30 @@ def make_get_request_to_webex(
     params,
     next_page_link,
     retry=True,
-    is_custom_endpoint=False
+    is_custom_endpoint=False,
+    webex_account_region="us_ca",
+    method = "GET",
+    payload=None
 ):
     if next_page_link:
         url = next_page_link
         params = None
-    else:
+    else: 
         url = _BASE_URL.format(base_endpoint=base_endpoint) + endpoint
+        protocol, rest = url.split("//")
         
         # reconstruct the url for meeting/qualities and cdr_feed endpoints
-        if not is_custom_endpoint and (endpoint == "meeting/qualities" or endpoint == "cdr_feed"):
-            protocol, rest = url.split("//")
+        if not is_custom_endpoint and endpoint == "meeting/qualities":
             url = f"{protocol}//analytics.{rest}"
+        elif not is_custom_endpoint and endpoint == "cdr_feed":
+            #construct the URL depending on the region
+            if webex_account_region == "us_ca":
+                url = f"{protocol}//analytics-calling.{rest}"
+            else:
+                url = f"{protocol}//analytics-calling-{webex_account_region}.{rest}"
        
-    helper.log_debug("[-] url: {} -- params: {}".format(url, params))
+    helper.log_debug("[-] url: {} -- method: {} -- params: {}".format(url, method, params))
+    
     headers = {
         "Authorization": "Bearer {access_token}".format(access_token=access_token),
     }
@@ -114,9 +131,9 @@ def make_get_request_to_webex(
         # use helper.send_http_request to have proxy enabled
         response = helper.send_http_request(
             url,
-            "GET",
+            method,
             parameters=params,
-            payload=None,
+            payload=payload,
             headers=headers,
             cookies=None,
             verify=False,
@@ -130,6 +147,7 @@ def make_get_request_to_webex(
                 response.status_code,
             )
         )
+        helper.log_debug(f"[-] Request method: {response.request.method}, Request body: {response.request.body}")
 
         data = None
         if response.status_code != 200:
@@ -182,7 +200,10 @@ def make_get_request_to_webex(
                     client_secret,
                     params,
                     retry=False,
-                    is_custom_endpoint=False
+                    is_custom_endpoint=False,
+                    webex_account_region="us_ca",
+                    method=method,
+                    payload=payload
                 )
             else:
                 response.raise_for_status()