+++ package/bin/webex_generic_endpoint.py
@@ -75,10 +75,24 @@ class ModInputWEBEX_GENERIC_ENDPOINT(base_mi.BaseModInput):
                 required_on_create=False,
             )
         )
-        
+
+        scheme.add_argument(
+            smi.Argument(
+                'method',
+                required_on_create=True,
+            )
+        )
+
         scheme.add_argument(
             smi.Argument(
                 'query_params',
+                required_on_create=False,
+            )
+        )
+
+        scheme.add_argument(
+            smi.Argument(
+                'request_body',
                 required_on_create=False,
             )
         )
@@ -87,7 +101,12 @@ class ModInputWEBEX_GENERIC_ENDPOINT(base_mi.BaseModInput):
 
     def validate_input(self, definition):
         """validate the input stanza"""
-        """Implement your own validation logic to validate the input stanza configurations"""
+        request_body = definition.parameters.get('request_body', None)
+        if request_body is not None:
+            try:
+                json.loads(request_body)
+            except json.JSONDecodeError:
+                raise ValueError("Invalid JSON string")
         pass
 
     def get_app_name(self):